+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # TBD
 * Small bugfix in `build_and_run.sh`, so it won't silently error on missing an argument
+* Update CI to skip byzantine tests on untrusted PRs where Docker Credentials are not present
 
 # 0.10.0
 * Upgraded to Kurtosis 1.0